Move `MockTimeProvider` into `TestHelpers`
After this change, all dsi-platform tests continue to complete successfully.
